Add a first attempt at moving the memoizing interpreter to the transformer
